Bail out early if we can tell that a parser actually ran
Since all of the parsers now have their own content-type header checks, it's safe to assume that if (for example) the request has "content-type: 'application/json'", then we can bail immediately after JSONBodyParser runs, because it definitely parsed the body (if there was an error, it would have been handled above).  It is possible that with some combination of options, the cascading down through several parsers still makes sense, so we're leaving it for now, but at some point we might just turn this into a big "switch" unless someone has a reasonable objection.